Api wallet dev

See merge request wallet/client/wallet_sdk!36

+++ wallet-api/src/infrastructure/task_queue/task/mod.rs
@@ -112,36 +112,38 @@ impl Tasks {
     async fn dispatch_tasks(
         entities: Vec<TaskQueueEntity>,
     ) -> Result<(), crate::error::service::ServiceError> {
-        let task_sender = crate::context::CONTEXT.get().unwrap().get_global_task_manager();
-        let pool = crate::context::CONTEXT.get().unwrap().get_global_sqlite_pool()?;
-        let mut repo = wallet_database::factory::RepositoryFactory::repo(pool.clone());
+        let handles = crate::context::CONTEXT.get().unwrap().get_global_handles();
+        if let Some(handles) = handles.upgrade() {
+            let task_sender = handles.get_global_task_manager();
+            let pool = crate::context::CONTEXT.get().unwrap().get_global_sqlite_pool()?;
+            let mut repo = wallet_database::factory::RepositoryFactory::repo(pool.clone());
 
-        let mut grouped_tasks: BTreeMap<u8, Vec<TaskQueueEntity>> = BTreeMap::new();
-        // let backend = crate::context::CONTEXT.get().unwrap().get_global_backend_api();
-        // let mut req = MsgAckReq::new();
-        for task_entity in entities.into_iter() {
-            // req.push(&task_entity.id);
-            match TryInto::<Box<dyn TaskTrait>>::try_into(&task_entity) {
-                Ok(task) => {
-                    let priority = super::task_manager::scheduler::assign_priority(&*task, false)?;
-                    grouped_tasks.entry(priority).or_default().push(task_entity);
-                }
-                Err(e) => {
-                    tracing::error!("task_entity.try_into() error: {}", e);
-                    repo.delete_task(&task_entity.id).await?;
-                }
-            };
-        }
-
-        for (priority, tasks) in grouped_tasks {
-            if let Err(e) = task_sender.get_task_sender().send(PriorityTask { priority, tasks }) {
-                tracing::error!("send task queue error: {}", e);
+            let mut grouped_tasks: BTreeMap<u8, Vec<TaskQueueEntity>> = BTreeMap::new();
+            // let backend = crate::context::CONTEXT.get().unwrap().get_global_backend_api();
+            // let mut req = MsgAckReq::new();
+            for task_entity in entities.into_iter() {
+                // req.push(&task_entity.id);
+                match TryInto::<Box<dyn TaskTrait>>::try_into(&task_entity) {
+                    Ok(task) => {
+                        let priority = super::task_manager::scheduler::assign_priority(&*task, false)?;
+                        grouped_tasks.entry(priority).or_default().push(task_entity);
+                    }
+                    Err(e) => {
+                        tracing::error!("task_entity.try_into() error: {}", e);
+                        repo.delete_task(&task_entity.id).await?;
+                    }
+                };
             }
+
+            for (priority, tasks) in grouped_tasks {
+                if let Err(e) = task_sender.get_task_sender().send(PriorityTask { priority, tasks }) {
+                    tracing::error!("send task queue error: {}", e);
+                }
+            }
+
+            repo.delete_oldest_by_status_when_exceeded(200000, 2).await?;
+            // backend.msg_ack(req).await?;
         }
-
-        repo.delete_oldest_by_status_when_exceeded(200000, 2).await?;
-        // backend.msg_ack(req).await?;
-
         Ok(())
     }
